Moved ip helpers into network/ip.js to not create another dir
